requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-snowflake==5.1.0
apache-airflow-providers-slack==8.3.0
dbt-snowflake==1.7.0
psycopg2-binary==2.9.9